Add new micro-task scheduler implementation
This patch adds a new implementation for the current global
micro-task-scheduler. It's backward compatible and works as
a drop-in replacement. The new implementation is struct based
and meant to support easier unit testing by being mocked out.

This patch is the first in a series of changes that refactor
portbase's module system to be less global-state heavy and
thus support easier and more powerfull testing of dependents.